requirements.txt
Django>=4.2,<5.0
django-import-export>=4.4,<5.0
djangorestframework>=3.17,<4.0
celery>=5.6,<6.0
redis>=7.4,<8.0
requests>=2.33,<3.0
